[python] move to common/ subdir
