+++ python/dune/xt/CMakeLists.txt
@@ -10,8 +10,3 @@
 # ~~~
 
 add_subdirectory(common)
-
-dune_pybindxi_add_module(_common EXCLUDE_FROM_ALL bindings.cc)
-dune_pybindxi_add_module(_empty EXCLUDE_FROM_ALL empty.cc)
-dune_pybindxi_add_module(_logging EXCLUDE_FROM_ALL logging.cc)
-dune_pybindxi_add_module(_timings EXCLUDE_FROM_ALL timings.cc)
