+++ python/dune/xt/common/CMakeLists.txt
@@ -6,8 +6,11 @@
 #      or  GPL-2.0+ (http://opensource.org/licenses/gpl-license)
 #          with "runtime exception" (http://www.dune-project.org/license.html)
 # Authors:
-#   René Fritze    (2018)
-#   Tobias Leibner (2018)
-#
-# do not delete
+#   René Fritze (2018)
 # ~~~
+
+dune_pybindxi_add_module(_common EXCLUDE_FROM_ALL bindings.cc)
+dune_pybindxi_add_module(_empty EXCLUDE_FROM_ALL empty.cc)
+dune_pybindxi_add_module(_logging EXCLUDE_FROM_ALL logging.cc)
+dune_pybindxi_add_module(_timings EXCLUDE_FROM_ALL timings.cc)
+
