+++ python/dune/xt/common/empty.cc
