--- python/dune/xt/empty/__init__.py
@@ -1,21 +1,0 @@
-# ~~~
-# This file is part of the dune-xt-common project:
-#   https://github.com/dune-community/dune-xt-common
-# Copyright 2009-2018 dune-xt-common developers and contributors. All rights reserved.
-# License: Dual licensed as BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
-#      or  GPL-2.0+ (http://opensource.org/licenses/gpl-license)
-#          with "runtime exception" (http://www.dune-project.org/license.html)
-# Authors:
-#   Felix Schindler (2017)
-#   René Fritze     (2018)
-# ~~~
-
-from importlib import import_module
-
-try:
-    from mpi4py import MPI
-except ImportError:
-    pass
-
-import dune.xt
-from dune.xt._empty import *
